Having explored last year's TechRadar by Devoteam, how do you envision evolving your IT ecosystem in 2024? Gain fresh insights with the launch of the updated 3rd edition of the TechRadar by Devoteam - on March 7th at 1 PM CET.
Join us for an online panel discussion with industry leaders from Devoteam and Veolia, broadcasted live on LinkedIn. This session promises to:

Highlight the top 150 technologies in TechRadar by Devoteam, focusing on cloud and data advancements that are making GenAI a reality.
Dive deep into how AI is reshaping daily life and IT ecosystems, with perspectives and case studies from Veolia and Devoteam experts.
Show practical strategies for using AI in a way that is both impactful and sustainable.
Are you ready to transform the hype of AI into value? Register now for the live session and be among the first to receive the latest TechRadar by Devoteam!